+++ plugins/nf-quilt/src/main/nextflow/quilt/QuiltProduct.groovy
@@ -17,6 +17,7 @@ package nextflow.quilt
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.ObjectWriter
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import nextflow.quilt.jep.QuiltPackage
 import nextflow.quilt.jep.QuiltParser
 import nextflow.quilt.nio.QuiltPath
@@ -96,7 +97,7 @@ ${nextflow}
         'homeDir', 'workDir', 'launchDir', 'manifest', 'configFiles'
     ]
 
-    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper()
+    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper().registerModule(new JavaTimeModule())
     private static final ObjectWriter OBJECT_WRITER = OBJECT_MAPPER.writerWithDefaultPrettyPrinter()
     static String toJson(Object value) {
         String result = OBJECT_WRITER.writeValueAsString(value)